project_name/requirements.txt
aniso8601==8.0.0
appdirs==1.4.3
atomicwrites==1.3.0
attrs==19.3.0
black==19.10b0
blinker==1.4
Click==7.0
colorama==0.4.3
Flask==1.1.1
flask-buzz==0.1.14
Flask-Mail==0.9.1
flask-marshmallow==0.10.1
flask-praetorian==1.0.0
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.4.1
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
marshmallow==3.3.0
marshmallow-sqlalchemy==0.21.0
more-itertools==8.1.0
packaging==20.0
passlib==1.7.2
pathspec==0.7.0
pendulum==2.0.5
pluggy==0.13.1
py==1.8.1
py-buzz==0.3.7
PyJWT==1.7.1
pyparsing==2.4.6
pytest==5.3.3
python-dateutil==2.8.1
python-dotenv==0.10.4
pytz==2019.3
pytzdata==2019.3
regex==2020.1.8
six==1.14.0
SQLAlchemy==1.3.12
toml==0.10.0
typed-ast==1.4.1
wcwidth==0.1.8
Werkzeug==0.16.0
